+++ packages/core/src/odata/common/selectable/binary-field.ts
@@ -5,10 +5,10 @@ import { EdmTypeShared } from '../edm-types';
 import { EntityBase, ODataVersionOf, Constructable } from '../entity';
 import {
   ComplexTypeField,
-  ConstructorOrField,
   getEdmType,
   getEntityConstructor
 } from './complex-type-field';
+import { ConstructorOrField } from './constructor-or-field';
 import { EdmTypeField, SelectableEdmTypeField } from './edm-type-field';
 
 /**