+++ packages/core/src/odata/common/selectable/complex-type-namespace.ts
@@ -10,6 +10,7 @@ export interface PropertyMetadata {
   name: string;
   originalName: string;
   type: EdmTypeShared<'any'> | ComplexTypeNamespace<any>;
+  isCollection?: boolean;
 }
 
 export function isComplexTypeNameSpace(