+++ test-packages/test-services/v2/test-service/TestComplexType.ts
@@ -205,67 +205,83 @@ export namespace TestComplexType {
   export const _propertyMetadata: PropertyMetadata[] = [{
     originalName: 'StringProperty',
     name: 'stringProperty',
-    type: 'Edm.String'
+    type: 'Edm.String',
+    isCollection: false
   }, {
     originalName: 'BooleanProperty',
     name: 'booleanProperty',
-    type: 'Edm.Boolean'
+    type: 'Edm.Boolean',
+    isCollection: false
   }, {
     originalName: 'GuidProperty',
     name: 'guidProperty',
-    type: 'Edm.Guid'
+    type: 'Edm.Guid',
+    isCollection: false
   }, {
     originalName: 'Int16Property',
     name: 'int16Property',
-    type: 'Edm.Int16'
+    type: 'Edm.Int16',
+    isCollection: false
   }, {
     originalName: 'Int32Property',
     name: 'int32Property',
-    type: 'Edm.Int32'
+    type: 'Edm.Int32',
+    isCollection: false
   }, {
     originalName: 'Int64Property',
     name: 'int64Property',
-    type: 'Edm.Int64'
+    type: 'Edm.Int64',
+    isCollection: false
   }, {
     originalName: 'DecimalProperty',
     name: 'decimalProperty',
-    type: 'Edm.Decimal'
+    type: 'Edm.Decimal',
+    isCollection: false
   }, {
     originalName: 'SingleProperty',
     name: 'singleProperty',
-    type: 'Edm.Single'
+    type: 'Edm.Single',
+    isCollection: false
   }, {
     originalName: 'DoubleProperty',
     name: 'doubleProperty',
-    type: 'Edm.Double'
+    type: 'Edm.Double',
+    isCollection: false
   }, {
     originalName: 'FloatProperty',
     name: 'floatProperty',
-    type: 'Edm.Float'
+    type: 'Edm.Float',
+    isCollection: false
   }, {
     originalName: 'TimeProperty',
     name: 'timeProperty',
-    type: 'Edm.Time'
+    type: 'Edm.Time',
+    isCollection: false
   }, {
     originalName: 'DateTimeProperty',
     name: 'dateTimeProperty',
-    type: 'Edm.DateTime'
+    type: 'Edm.DateTime',
+    isCollection: false
   }, {
     originalName: 'DateTimeOffSetProperty',
     name: 'dateTimeOffSetProperty',
-    type: 'Edm.DateTimeOffset'
+    type: 'Edm.DateTimeOffset',
+    isCollection: false
   }, {
     originalName: 'ByteProperty',
     name: 'byteProperty',
-    type: 'Edm.Byte'
+    type: 'Edm.Byte',
+    isCollection: false
   }, {
     originalName: 'SByteProperty',
     name: 'sByteProperty',
-    type: 'Edm.SByte'
+    type: 'Edm.SByte',
+    isCollection: false
   }, {
     originalName: 'ComplexTypeProperty',
     name: 'complexTypeProperty',
-    type: TestNestedComplexType
+    type: TestNestedComplexType,
+    isCollection: false
   }];
   /**
    * Type reference to the according complex type.