+++ test-packages/test-services/v4/test-service/TestLvl2NestedComplexType.js
@@ -59,7 +59,8 @@ var TestLvl2NestedComplexType;
     TestLvl2NestedComplexType._propertyMetadata = [{
             originalName: 'StringProperty',
             name: 'stringProperty',
-            type: 'Edm.String'
+            type: 'Edm.String',
+            isCollection: false
         }];
     /**
      * Type reference to the according complex type.